Add support for custom certificates
The parameters are added to the provider config to support it:
- ca_cert
- client_cert
- client_key

You need to have all three passed for the configuration to be picked up,
it will fail otherise. It also overrides the `tls` parameter
(cfg.TLSConfig), if the latter is set.

Tests are not passing (they were passing upstream either), but running
it locally works fine.